Add core of `openRoom` to a component effect
