+++ apps/meteor/client/views/root/AppLayout.tsx
@@ -3,11 +3,13 @@ import { useSyncExternalStore } from 'use-sync-external-store/shim';
 
 import { appLayout } from '../../lib/appLayout';
 import { blazePortals } from '../../lib/portals/blazePortals';
+import { useOpenRoomHandling } from '../room/hooks/useOpenRoomHandling';
 import PageLoading from './PageLoading';
 import { useTooltipHandling } from './useTooltipHandling';
 
 const AppLayout: FC = () => {
 	useTooltipHandling();
+	useOpenRoomHandling();
 
 	const layout = useSyncExternalStore(appLayout.subscribe, appLayout.getSnapshot);
 	const portals = useSyncExternalStore(blazePortals.subscribe, blazePortals.getSnapshot);
